+++ components/feedback-modal.tsx
@@ -67,7 +67,8 @@ export function FeedbackModal({ open, onOpenChange }: FeedbackModalProps) {
         <DialogHeader>
           <DialogTitle>Give feedback</DialogTitle>
           <DialogDescription>
-            Your feedback helps us improve Vana. Let us know what you think!
+            Your feedback helps us improve Polymorph. Let us know what you
+            think!
           </DialogDescription>
         </DialogHeader>